+++ src/models/ntlm.sso.config.model.ts
@@ -1,0 +1,3 @@
+export interface NtlmSsoConfig {
+  ntlmHosts: string[];
+}