+++ src/proxy/config.controller.ts
@@ -8,6 +8,8 @@ import { IConfigController } from './interfaces/i.config.controller';
 import { IConfigStore } from './interfaces/i.config.store';
 import { TYPES } from './dependency.injection.types';
 import { IDebugLogger } from '../util/interfaces/i.debug.logger';
+import { SsoConfigValidator } from '../util/sso.config.validator';
+import { NtlmSsoConfig } from '../models/ntlm.sso.config.model';
 
 @injectable()
 export class ConfigController implements IConfigController {
@@ -22,6 +24,7 @@ export class ConfigController implements IConfigController {
     this._configStore = configStore;
     this._debug = debug;
     this.router.post('/ntlm-config', (req: Request, res: Response) => this.ntlmConfig(req, res));
+    this.router.post('/ntlm-sso', (req: Request, res: Response) => this.ntlmSso(req, res));
     this.router.post('/reset', (req: Request, res: Response) => this.reset(req, res));
     this.router.get('/alive', (req: Request, res: Response) => this.alive(req, res));
     this.router.post('/quit', (req: Request, res: Response) => this.quit(req, res));
@@ -29,22 +32,34 @@ export class ConfigController implements IConfigController {
 
   private ntlmConfig(req: Request, res: Response) {
     let validateResult = ConfigValidator.validate(req.body);
-      if (!validateResult.ok) {
-        res.status(400).send('Config parse error. ' + validateResult.message);
+    if (!validateResult.ok) {
+      res.status(400).send('Config parse error. ' + validateResult.message);
+    } else {
+      this._debug.log('Received valid NTLM config update');
+      let config = req.body as NtlmConfig;
+      let ntlmHostUrl = toCompleteUrl(config.ntlmHost, false);
+      if (this._configStore.exists(ntlmHostUrl)) {
+        // Trigger removal of existing authentication cache
+        this.configApiEvent.emit('configUpdate', ntlmHostUrl);
+        this._debug.log('Updating host', ntlmHostUrl.href);
       } else {
-        this._debug.log('Received valid config update');
-        let config = req.body as NtlmConfig;
-        let ntlmHostUrl = toCompleteUrl(config.ntlmHost, false);
-        if (this._configStore.exists(ntlmHostUrl)) {
-          // Trigger removal of existing authentication cache
-          this.configApiEvent.emit('configUpdate', ntlmHostUrl);
-          this._debug.log('Updating host', ntlmHostUrl.href);
-        } else {
-          this._debug.log('Added new host' , ntlmHostUrl.href);
-        }
-        this._configStore.updateConfig(config);
-        res.sendStatus(200);
+        this._debug.log('Added new host' , ntlmHostUrl.href);
       }
+      this._configStore.updateConfig(config);
+      res.sendStatus(200);
+    }
+  }
+
+  private ntlmSso(req: Request, res: Response) {
+    let validateResult = SsoConfigValidator.validate(req.body);
+    if (!validateResult.ok) {
+      res.status(400).send('SSO config parse error. ' + validateResult.message);
+    } else {
+      this._debug.log('Received valid NTLM SSO config');
+      let config = req.body as NtlmSsoConfig;
+      this._configStore.setSsoConfig(config);
+      res.sendStatus(200);
+    }
   }
 
   private reset(req: Request, res: Response) {