+++ src/proxy/config.store.ts
@@ -3,6 +3,7 @@ import { toCompleteUrl } from '../util/url.converter';
 import { CompleteUrl } from '../models/complete.url.model';
 import { injectable } from 'inversify';
 import { IConfigStore } from './interfaces/i.config.store';
+import { NtlmSsoConfig } from '../models/ntlm.sso.config.model';
 
 interface NtlmHostConfigHash {
   [ntlmHost: string]: NtlmConfig;
@@ -11,6 +12,7 @@ interface NtlmHostConfigHash {
 @injectable()
 export class ConfigStore implements IConfigStore {
   private ntlmHosts: NtlmHostConfigHash = {};
+  private ntlmSsoHosts: string[] = [];
 
   updateConfig(config: NtlmConfig) {
     let ntlmHostUrl = toCompleteUrl(config.ntlmHost, false);
@@ -34,7 +36,23 @@ export class ConfigStore implements IConfigStore {
     return this.ntlmHosts[ntlmHostUrl.href];
   }
 
+  setSsoConfig(ntlmSsoConfig: NtlmSsoConfig) {
+    this.ntlmSsoHosts = ntlmSsoConfig.ntlmHosts;
+  }
+
+  useSso(ntlmHostUrl: CompleteUrl): boolean {
+    if (this.ntlmSsoHosts.includes(ntlmHostUrl.hostname) && this.exists(ntlmHostUrl) === false) {
+      return true;
+    }
+    return false;
+  }
+
+  existsOrUseSso(ntlmHostUrl: CompleteUrl): boolean {
+    return this.exists(ntlmHostUrl) || this.ntlmSsoHosts.includes(ntlmHostUrl.hostname);
+  }
+
   clear() {
     this.ntlmHosts = {};
+    this.ntlmSsoHosts = [];
   }
 }